+++ src/main/java/com/bway/ecommerce/controller/CategoryController.java
@@ -2,6 +2,8 @@ package com.bway.ecommerce.controller;
 
 
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -23,38 +25,53 @@ public class CategoryController {
 	private CategoryService catService;
 	
 	@GetMapping("/add")
-	public String getCategory() {
+	public String getCategory(HttpSession session) {
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		} 
 		return "categoryadd";
 	}
 	@PostMapping("/add")
-	public String postCategory(@ModelAttribute Category cat) {
+	public String postCategory(HttpSession session ,@ModelAttribute Category cat) {
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		} 
 		catService.addCategory(cat);
 		return "categoryadd";
 	}
 	@GetMapping("/list")
-	public String catList(Model model) {
+	public String catList(Model model, HttpSession session) {
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		} 
 		model.addAttribute("catList", catService.getAllCats());
 		return "categorylist";
 	}
 	
 	@GetMapping("/edit")
-	public String edit(@RequestParam long id, Model model){
+	public String edit(HttpSession session,@RequestParam long id, Model model){
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		} 
 		model.addAttribute("catObject", catService.getCatById(id));
 		return "categoryedit";
 	}
 	@GetMapping("/delete")
-	public String delete(@RequestParam long id) {
+	public String delete(HttpSession session, @RequestParam long id) {
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		} 
 		catService.deleteCat(id);
 		return "redirect:/category/list";
 	}
 	@PostMapping("/update")
-	public String update(@ModelAttribute Category cat) {
-		 catService.updateCat(cat);
+	public String update(@ModelAttribute Category cat, HttpSession session) {
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		}  
+		catService.updateCat(cat);
 		 return "redirect:/category/list";
 	}
 	
-	@GetMapping("/table")
-	public String getTable() {
-		return "tables";
-	}
+	
 }
