+++ src/main/java/com/bway/ecommerce/controller/ProductController.java
@@ -1,5 +1,7 @@
 package com.bway.ecommerce.controller;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -30,15 +32,20 @@ public class ProductController {
 	private BrandService brandService;
 	
 	@GetMapping("/add")
-	public String getProduct(Model model) {
+	public String getProduct(Model model, HttpSession session) {
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		}  
 		model.addAttribute("catList", catService.getAllCats());
 		model.addAttribute("brandList", brandService.getAllBrands());
 		return "productadd";
 	}
 	
 	@PostMapping("/add")
-	public String postProduct(@ModelAttribute Product product, @RequestParam MultipartFile photo) {
-		
+	public String postProduct(HttpSession session ,@ModelAttribute Product product, @RequestParam MultipartFile photo) {
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		}  
 		if(!photo.isEmpty()) {
 			product.setImage(photo.getOriginalFilename());
 			fileUtil.fileUploadProduct(photo);
@@ -47,25 +54,37 @@ public class ProductController {
 		return "redirect:/product/add";
 	}
 	@GetMapping("/list")
-	public String getAll(Model model) {
+	public String getAll(Model model, HttpSession session) {
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		}  
 		model.addAttribute("productList",productService.getAllProducts());
 		return "productlist";
 	}
 	@GetMapping("/delete")
-	public String delete(@RequestParam Long id) {
+	public String delete(@RequestParam Long id, HttpSession session) {
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		}  
 		productService.deleteProduct(id);
 		return "redirect:/product/list";
 	}
 	@GetMapping("/edit") 
-	public String edit(@RequestParam Long id, Model model) {
+	public String edit(@RequestParam Long id, Model model, HttpSession session) {
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		}  
 		model.addAttribute("productObject", productService.getProductById(id));
 		model.addAttribute("catList", catService.getAllCats());
 		model.addAttribute("brandList", brandService.getAllBrands());
 		return "productedit";
 	}
 	@PostMapping("/update")
-	public String update(@ModelAttribute Product product) {
-		 productService.updateProduct(product);
+	public String update(HttpSession session, @ModelAttribute Product product) {
+		if(session.getAttribute("validuser") == null) {
+			return "login";
+		}   
+		productService.updateProduct(product);
 		 return "redirect:/product/list";
 	}
 }
